+++ server/Recruit-Api/src/main/java/com/econovation/recruit/api/recruitment/util/LatestRecruitmentVo.java
@@ -1,7 +1,7 @@
 package com.econovation.recruit.api.recruitment.util;
 
 import com.econovation.recruitdomain.common.aop.redissonLock.RedissonLock;
-import com.econovation.recruitdomain.domains.applicant.domain.state.RecruitmentStates;
+import com.econovation.recruitdomain.domains.recruitment.domain.RecruitmentStates;
 import com.econovation.recruitdomain.domains.recruitment.domain.Recruitment;
 import java.time.LocalDateTime;
 import java.util.Objects;
