+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/dto/RecruitmentResponseDto.java
@@ -1,6 +1,6 @@
 package com.econovation.recruitdomain.domains.dto;
 
-import com.econovation.recruitdomain.domains.applicant.domain.state.RecruitmentStates;
+import com.econovation.recruitdomain.domains.recruitment.domain.RecruitmentStates;
 import com.econovation.recruitdomain.domains.recruitment.domain.Recruitment;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
