+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/recruitment/adaptor/RecruitmentAdaptor.java
@@ -1,7 +1,7 @@
 package com.econovation.recruitdomain.domains.recruitment.adaptor;
 
 import com.econovation.recruitcommon.annotation.Adaptor;
-import com.econovation.recruitdomain.domains.applicant.domain.state.RecruitmentStates;
+import com.econovation.recruitdomain.domains.recruitment.domain.RecruitmentStates;
 import com.econovation.recruitdomain.domains.recruitment.domain.Recruitment;
 import com.econovation.recruitdomain.domains.recruitment.domain.RecruitmentRepository;
 import com.econovation.recruitdomain.out.RecruitmentPort;
