+++ server/Recruit-Domain/src/main/java/com/econovation/recruitdomain/domains/recruitment/domain/RecruitmentStates.java
@@ -1,4 +1,4 @@
-package com.econovation.recruitdomain.domains.applicant.domain.state;
+package com.econovation.recruitdomain.domains.recruitment.domain;
 
 public enum RecruitmentStates {
     NON_START,
